+++ CHANGELOG.md
@@ -9,9 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added a Github Action template for Maven build
 - Added Apache Maven support
 - Improving Pipeline to reduce the maintenance
 - Added DevContainer support
-- Added option to show a hint about Spring CLI usage
+- Added Spring CLI support
 - Cursor rules update. Now all cursor rules has the extension .mdc
 - Added CHANGELOG.md